Improve phpstan typechecks
There are still plenty issues with typing with phpstan generics:
1) Generics in closures does not work properly - type inference for
generics there does not work in phpstan and it complains. On many
places I was forced to use mixed :-(

2) General iterable typechecks does not work well in all cases

3) Scalar union types cannot be used as upper bounds for generics

4) Array tuples generics does not work well in all cases

5) There is no (simple) way to tell which generics to use in
call context. This is a bit issue for calls, where in proper generics
you would parametrize generic explicitly (because it cannot be infered).
E.g. ArrayList::fromEmpty() - (there is no context from which you can
infer generic type and with proper generics you would parametrize it
explicitly).