Improve typesafety with correct method use
When of is called in variadic way it infers type of collection from the
very first item passed into it. If we need more wide type (like union of
int|string) we need to create that collection different way. This is
where fromIterable servers better for this case, becuase passed
collection (array in many cases) may has already type widened by its
contents.

Also, when of is called without any argument, the structure type is
infered to mixed, this is where fromEmpty servers better. Even though
sometimes it is still needed to annotate it's type explicitelly